Simplified directory reading.
No longer refresh the list of files upon successful processing of a batch
job.

This makes in much harder to accidentally create infinite loops, since you
no longer have to keep in mind that the directory is rescanned (and the
processing of batches restarted) every time a callback returns true.

The drawback is that if a finishing batch job creates new files (or renames
old ones) the files with these new names will never be picked up by the
batch job processor. (You'll have to manually run `batch` again to process
these.)